+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--
+- More resilient RequirementCache that checks for module import-ability ([#112](https://github.com/Lightning-AI/utilities/pull/112))
 
 
 ## [0.7.0] - 2023-02-20